requirements.txt
scikit-learn==1.3.2
xgboost==3.0.2
pandas
numpy
joblib
fastapi==0.78.0
uvicorn==0.18.3
pymongo==4.2.0
python-dotenv==0.21.0
python-multipart
dnspython==2.2.1
